+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Added
 
+- added `get_big_lay_nice` function <2025-05-23, Fri>
 - added `group_zoom` argument for `g_layout_nice` function <2025-05-23, Fri>
 - added `plot_e_type_bar`, `plot_multi_nets` functions <2025-05-22, Thu>
 - added `c_net_ego`, `c_net_intersect`, `c_net_difference`, `c_net_highlight` functions <2025-04-28, Mon>